Add ability to ignore some points
Any points in IgnorePoints will not be used in the triangulation.

Use case: I have a mesh, and that mesh's vertex data needs to stay the same while its triangles are regenerated to ignore a varying set of vertices.

This is awkward and WIP, and the triangulator now breaks if you don't supply it with a HashSet of points to ignore.

I don't really know what I'm doing with Burst and haven't used it before. Need to revisit this later and make it better.